* Move shared functions to utils
